requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
numpy==1.22.4
Pillow==9.1.1
pytz==2022.1
qrcode==7.3.1
rectangle==0.4
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
yarl==1.7.2
